% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_time_bins.R
\name{add_time_bins}
\alias{add_time_bins}
\title{Add Time Bins Relative to Noun Onset}
\usage{
add_time_bins(df_et)
}
\arguments{
\item{df_et}{A data frame of eye tracking data}
}
\description{
\code{add_time_bins()} is used to add time bins to Tobii eyetracking data
}
\details{
This function allows you to add time bins for Tobii eyetracking data. This function ensures
that the timestamps will be similar across participants and trials.
}
\examples{
\dontrun{d <- add_time_bins(df_et)}
}
