% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_time_crit_onset.R
\name{add_time_crit_onset}
\alias{add_time_crit_onset}
\title{Add Critical Onset Values And Compute Relative Time In Trial}
\usage{
add_time_crit_onset(df_et, df_timing, onset_type = "noun")
}
\arguments{
\item{df_et}{A data frame of eye tracking data}

\item{df_timing}{A data frame of trial timing information}

\item{onset_type}{A string indicating the type of onset to compute ("sound" or "noun").}
}
\description{
\code{add_time_crit_onset()} is used to add time relative to critical points
 in each trial. The two critical points include: sound and noun.
}
\details{
This function allows you to add critical onsets (sound on, noun on) to Tobii eyetracking data
based on timing coded in an external timing file.
}
\examples{
\dontrun{d <- add_time_crit_onset(df_et, d_timing, onset_type = "noun")}
}
