% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_aoi_looking.R
\name{code_aoi_looking}
\alias{code_aoi_looking}
\title{Code AOI Looking}
\usage{
code_aoi_looking(df_et, aois_list)
}
\arguments{
\item{df_et}{A data frame of eye tracking data}

\item{aois_list}{A names list of AOI values of the following format.
\code{aois_list <- list(l_xmin=7, l_xmax=396, l_ymin=1749, l_ymax=2044, r_xmin=1289, r_xmax=1685, l_ymin=1751, l_ymax=2044)}}
}
\description{
\code{code_aoi_looking()} is used to code looking behavior based on user-specified AOIs for
 Tobii eyetracking data
}
\details{
This function scores each look as left, right, away based on gaze coordinates (gaze_x and gaze_y_cartesian).
}
\examples{
\dontrun{d <- code_aoi_looking(df_et, aois_list = aois)}
}
