% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_target_looking.R
\name{code_target_looking}
\alias{code_target_looking}
\title{Code Target Looking}
\usage{
code_target_looking(df_et)
}
\arguments{
\item{df_et}{A data frame of eye tracking data with looks that have been
coded as left, right, away by \code{code_aoi_looking()} function.}
}
\description{
\code{code_target_looking()} is used to add 1, 0, 0.5 codes (hits, misses, aways) to looking behavior bins for Tobii eyetracking data
}
\details{
This function scores each look as left, right, away based on gaze coordinates (gaze_x and gaze_y_cartesian).
}
\examples{
\dontrun{d <- code_target_looking(df_et, aois_list = aois)}
}
