% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_wide.R
\name{convert_to_wide}
\alias{convert_to_wide}
\title{Convert From Long To Wide Format}
\usage{
convert_to_wide(df_et_long)
}
\arguments{
\item{df_et_long}{A data frame of eye tracking data in long format}
}
\description{
\code{convert_to_wide()} is used to convert processed Tobii data from long to wide format
}
\details{
This function converts Tobii eye tracking data from long to wide format where time bins become columns and each
row is a trial.
}
\examples{
\dontrun{d <- convert_to_wide(df_et_long)}
}
