% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_responses.R
\name{identify_responses}
\alias{identify_responses}
\title{Code Responses For Each Trial}
\usage{
identify_responses(df_et)
}
\arguments{
\item{df_et}{A data frame of eye tracking data}
}
\description{
\code{identify_responses()} is used to code looking behavior based on user-specified AOIs for
 Tobii eyetracking data
}
\details{
This function identifies each trial "response" as D, T or A, based on "hit" or "miss" or "off" at F0.
}
\examples{
\dontrun{d <- identify_responses(df_et)}
}
