% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_iChart_names.R
\name{make_iChart_names}
\alias{make_iChart_names}
\title{Make iChart Names Match iChartAnalyzeR formatting}
\usage{
make_iChart_names(df)
}
\arguments{
\item{df}{A data frame of eye tracking data in wide format}
}
\description{
\code{make_iChart_names()} is used to make column names that work for iChartAnalyzeR functions
}
\details{
This function changes the column names in the Tobii eye tracking data to match those in iChartAnalyzeR functions.
}
\examples{
\dontrun{d <- make_iChart_names(df)}
}
