% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fixations.R
\name{plot_fixations}
\alias{plot_fixations}
\title{Plot Fixations With AOIs for PEEK Tobii Data}
\usage{
plot_fixations(et_df, frac_size = 0.05, aois, ...)
}
\arguments{
\item{et_df}{A data frame of Tobii data, usually created by the \code{read_all_tsvs()} function.}

\item{frac_size}{A double indicating the fraction of rows (fixations) to plot for each participant.
Useful for controlling how long the plot will take to render.}

\item{aois}{An AOI data frame indicating the AOI coordinates, usually this object is created
by the \code{read_aois()} function.}

\item{...}{Additional arguments, typically used to control the \code{read_delim() function}.}
}
\description{
\code{plot_fixations()} is a \code{ggplot()}.
   It's useful for doing a quick check that your AOIs look reasonable when plotted
   with the raw eyetracking data.
}
\details{
This function allows you to make an exploratory plot of the Tobii .tsv eyetracking data with
the AOIs read from the AOI .txt file.
}
\examples{
\dontrun{plot_fixations(tsv_data, frac_size = .05, aois)}
}
