% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_timecourses.R
\name{plot_timecourses}
\alias{plot_timecourses}
\title{Plot Timecourses of Raw Looking Behavior For PEEK Tobii Data}
\usage{
plot_timecourses(df_et, n_subs = 3, n_trials = 3)
}
\arguments{
\item{df_et}{A data frame of Tobii data, usually created by the \code{read_tobii()} function.}

\item{n_subs}{An integer indicating the number of subjects to include in the plot}

\item{n_trials}{An integer indicating the number of trials to include in the plot}

\item{...}{Additional arguments.}
}
\description{
\code{plot_timecourses()} is a \code{ggplot()}.
   It's useful for doing a quick check that your timecourse looking data appear reasonable
}
\details{
This function allows you to make an exploratory plot of the timecourse looking from
Tobii .tsv eyetracking data
}
\examples{
\dontrun{plot_timecourses(tsv_data, n_subs = 3, n_trials = 3)}
}
