% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_all_tsvs.R
\name{read_all_tsvs}
\alias{read_all_tsvs}
\title{Read And Parse All .tsv Files In Directory}
\usage{
read_all_tsvs(dir_path, tidy = FALSE, y_max = 2048)
}
\arguments{
\item{dir_path}{A path to a directory of .tsv files.}

\item{tidy}{A boolean, if tidy = FALSE this functions behaves just \code{read_tsv()}.}

\item{y_max}{An integer value indicating the largest y-coordinate in the data set.
Used to create a cartesian y-coordinate variable.}
}
\description{
\code{read_all_tsvs()} reads parses a directory of .vcx coding files to
  to get prescreening information.
}
\details{
This function allows you to read and parse all of the raw vcx files generated by iCoder.
}
\examples{
  \dontrun{d <- read_all_tsvs(dir_path = "experiment_name/data/tsv_files/")}
}
