% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_all_vcx.R
\name{read_all_vcx}
\alias{read_all_vcx}
\title{Read And Parse All .vcx Files In Directory}
\usage{
read_all_vcx(dir_path)
}
\arguments{
\item{dir_path}{A path to a directory of .vcx files.}
}
\description{
\code{read_all_vcx()} reads parses a directory of .vcx coding files to
  to get prescreening information. The
}
\details{
This function allows you to read and parse all of the raw vcx files generated by iCoder.
}
\examples{
  d <- read_all_vcx(dir_path = "experiment_name/data/vcx_files/")
}
