% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_aois.R
\name{read_aois}
\alias{read_aois}
\title{Read And Tidy PEEK AOIS File}
\usage{
read_aois(file, tidy = FALSE, y_max = 2048, ...)
}
\arguments{
\item{file}{Either a path to a file, a connection, or literal data.}

\item{tidy}{A boolean, if tidy = FALSE this functions behaves just \code{read_delim()}.}

\item{y_max}{An integer value indicating the largest y-coordinate in the data set. Used to create a}

\item{...}{Additional arguments, typically used to control the \code{read_delim() function}.}
}
\description{
\code{read_aois()} is a special case of \code{read_delm()}.
   It's useful for doing minor cleaning of the AOI coordinates. If you set \code{tidy = FALSE},
   the function behaves like the default \code{read_delim()}. Note that
   you are able to pass along any arguments that you would use with \code{read_delim()}.
}
\details{
This function allows you to read and parse a single PEEK order file generated by the Language
Learning Lab
}
\examples{
\dontrun{d <- read_aois(file = "Habla2_25_Clips_aoi.txt", tidy = TRUE, y_max = 2048)}
}
