% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_order_file.R
\name{read_order_file}
\alias{read_order_file}
\title{Read And Parse PEEK Order File}
\usage{
read_order_file(file, tidy = TRUE, ...)
}
\arguments{
\item{file}{Either a path to a file, a connection, or literal data.}

\item{tidy}{A boolean, if tidy = FALSE this functions behaves just \code{read_delim()}.}

\item{...}{Additional arguments, typically used to control the \code{read_delim() function}.}
}
\description{
\code{read_order_file()} is a special case of \code{read_delm()}.
   It's useful for doing minor cleaning of variable names. If you set \code{tidy = FALSE},
   the function behaves like the default \code{read_delim()}. Note that
   you are able to pass along any arguments that you would use with \code{read_tsv()}.
}
\details{
This function allows you to read and parse a single PEEK order file generated by the Language
Learning Lab
}
\examples{
\dontrun{d <- read_order_file(file = "Habla2_25_Clips_order.txt", tidy = TRUE)}
}
