% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_timing_file.R
\name{read_timing_file}
\alias{read_timing_file}
\title{Read And Parse PEEK Trial Timing File}
\usage{
read_timing_file(file, tidy = TRUE, fps = 29.97, ...)
}
\arguments{
\item{file}{Either a path to a file, a connection, or literal data.}

\item{tidy}{A boolean, if tidy = FALSE this functions behaves just \code{read_delim()}.}

\item{...}{Additional arguments, typically used to control the \code{read_tsv() function}.}
}
\description{
\code{read_timing_file()} is a special case of \code{read_delm()}.
   It's useful for converting the \code{00:01:22} (min, sec, frames) format
   to milliseconds for each key event in the trial (e.g., Sound On, Noun On, etc.)
   If you set \code{tidy = FALSE}, the function behaves like the default \code{read_delim()}.
   Note that you are able to pass along any arguments that you would use with \code{read_delim()}.
}
\details{
This function allows you to read and parse a single raw timing file generated
by human coders marking key points in the trial
}
\examples{
\dontrun{d <- read_order_file(file = "Habla2_25_Clips_order.txt", tidy = TRUE)}
}
