% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tobii.R
\name{read_tobii}
\alias{read_tobii}
\title{Read And Parse Tobii File}
\usage{
read_tobii(file, tidy = FALSE, ...)
}
\arguments{
\item{file}{Either a path to a file, a connection, or literal data.}

\item{tidy}{A boolean, if tidy = FALSE this functions behaves just \code{read_tsv()}.}

\item{...}{Additional arguments, typically used to control the \code{read_tsv() function}.}
}
\description{
\code{read_tobii()} is a special case of \code{read_tsv()}. It's useful for doing minor data
   processing on the raw tsv file that is unique to the Language Learning Lab.
   If you set \code{tidy = FALSE}, the function behaves like the default \code{read_tsv()}. Note that
   you are able to pass along any arguments that you would use with \code{read_tsv()}.
}
\details{
This function allows you to read and parse a single raw tsv file generated by Tobii
}
\examples{
\dontrun{d <- read_tobii(file = "187_Habla2_25_Clips.tsv", tidy = TRUE)}
}
