% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_vcx.R
\name{read_vcx}
\alias{read_vcx}
\title{Read And Parse .vcx Files}
\usage{
read_vcx(file)
}
\arguments{
\item{file}{Either a path to a file, a connection, or literal data.}
}
\description{
\code{read_vcx()} reads parses raw .vcx coding files to get prescreening information
  to filter unusable trials from eyetracking data. The bulk of the function is
  doing extraction of strings from deeply nested XML data structure.
}
\details{
This function allows you to read and parse raw vcx files generated by iCoder
}
\examples{
  \dontrun{d <- read_vcx(file = "187_Habla2_25_Clips.vcx")}
}
