% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fixations.R
\name{plot_fixations}
\alias{plot_fixations}
\title{Plot Fixations With AOIs for PEEK Tobii Data}
\usage{
plot_fixations(df_et, frac_size = 0.05, aois_list, ...)
}
\arguments{
\item{df_et}{A data frame of Tobii data, usually created by the \code{read_tobii()} function.}

\item{frac_size}{A double indicating the fraction of rows (fixations) to plot for each participant.
Useful for controlling how long the plot will take to render.}

\item{aois_list}{A named list of AOI coordinates with the structure:
\code{aois_list <- list(l_xmin=7, l_xmax=396, l_ymin=1749, l_ymax=2044, r_xmin=1289, r_xmax=1685, l_ymin=1751, l_ymax=2044)}}

\item{...}{Additional arguments, typically used to pass along to ggplot.}
}
\description{
\code{plot_fixations()} is a \code{ggplot()}.
   It's useful for doing a quick check that your AOIs look reasonable when plotted
   with the raw eyetracking data.
}
\details{
This function allows you to make an exploratory plot of the Tobii .tsv eyetracking data with
the AOIs read from the AOI .txt file.
}
\examples{
\dontrun{plot_fixations(df_et, frac_size = .05, aois_list)}
}
